Deprecated functions in package eixport.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("-deprecated").

The Emissions Database for Global Atmospheric Research (EDGAR) is a project from the Joint Research Centre. They provide provides global past and present day anthropogenic emissions of greenhouse gases and air pollutants by country and on spatial grid. get_edgar provide functions to download any of the EDGAR datasets.

get_edgar(
  dataset = "v50_AP",
  pol,
  sector,
  year,
  n = parallel::detectCores()/2,
  destpath = tempdir(),
  type = "nc",
  ask = TRUE,
  copyright = TRUE,
  verbose = TRUE
)

Arguments

dataset

Character; name of the datasets: "v50_AP", "v432_AP", "v432_VOC_spec", "htap_v2_2", "v50_GHG

pol

Character; one of the pollutants shown on note.

sector

Character; one of the sectors shown on note.

year

Integer; years.

n

Integer; Number of cores in the machine, by default the half.

destpath

Character: Path to create the directory for downloads datasets

type

Extension, character/ Indicates if the file should be "txt", "nc" or "other". txt with untis t/year and nc with units are ug/m2/s

ask

Logical; Are these URL ok?

copyright

Logical; to show copyright information.

verbose

Logical, to print more information.

Value

Downloads data

Note

I recommend 2 ways:

1. include 'sector' and do not include 'pol', which download all pollutants as default

get_edgar(dataset = "v432_AP", destpath = tempdir(), sector = c("TRO", "TOTALS"), year = 2012)

2. include 'pol' and do not include 'sector', which download all sectors as default

get_edgar(dataset = "v432_AP", destpath = tempdir(), pol = c("CO", "NOx"), year = 2012)

dataset	pollutant	sector	years

v50_AP	BC, CO, NH3, NMVOC, NOx, OC, PM10, PM2.5, SO2	"AWB" "CHE" "ENE" "FFF" "FOO_PAP" "IND" "IRO" "NFE" "NMM" "PRO" "RCO" "REF_TRF" "SWD_INC" "TNR_Aviation_CDS" "TNR_Aviation_CRS" "TNR_Aviation_LTO" "TNR_Aviation_SPS" "TNR_Other" "TNR_Ship" "TOTALS" "TRO_RES" "TRO_noRES" "AGS" "MNM" "PRU_SOL" "SWD_LDF" "WWT" "NEU"	1970-2015

v50_GHG	CH4,CO2_excl_short-cycle_org_C, CO2_org_short-cycle_C, N2O	AGS AWB CHE ENE ENF FFF IND IRO MNM PRO PRO_COAL PRO_GAS PRO_OIL RCO REF_TRF SWD_INC SWD_LDF TNR_Aviation_CDS TNR_Aviation_CRS TNR_Aviation_LTO TNR_Aviation_SPS TNR_Other TNR_Ship TOTALS TRO WWT NEU NFE NMM PRU_SOL IDE N2O	1970-2018

v432	CH4,CO2_excl_short-cycle_org_C, CO2_org_short-cycle_C, N2O	AGS AWB CHE ENE ENF FFF IND IRO MNM PRO RCO REF_TRF SWD_INC SWD_LDF TNR_Aviation_CDS TNR_Aviation_CRS TNR_Aviation_LTO TNR_Aviation_SPS TNR_Other TNR_Ship TOTALS TRO WWT NEU NFE NMM PRU_SOL IDE N2O	1970-2012

v432_AP	BC, CO, NH3, NMVOC, NOx, OC, PM10, PM2.5_bio, PM2.5_fossil, SO2	AWB CHE ENE FFF FOO_PAP IND IRO NFE NMM PRO RCO REF_TRF SWD_INC TNR_Aviation_CDS TNR_Aviation_CRS TNR_Aviation_LTO TNR_Aviation_SPS TNR_Other TNR_Ship TOTALS TRO AGS MNM PRU_SOL SWD_LDF NEU	1970-2012

v432_VOC_spec	voc1, voc2, voc3, voc3, voc5, voc6, voc7, voc8, bvoc9, voc10. voc11, voc12, voc13, voc14, voc15, voc16, voc17, voc18, voc19, voc20, voc21, voc22, voc23, voc24, voc25, TOTALS	ENE, IND, PPA, RCO, REF TNR_Aviation_CDS, TNR_Aviation_CRS, TNR_Aviation_LTO, TNR_Aviation_SPS, TNR_Other, TNR_Ship, TOTALS, TRO, TRF	1970-2012

htap_v2_2	BC, CO, NH3, NMVOC, NOx, OCm, PM10, PM2.5, SO2	AIR, ENERGY, INDUSTRY, RESIDENTIAL, SHIPS TRANSPORT	2008 and 2010 \cr

voc11 only 2008

References

v50_AP Crippa, M., Solazzo, E., Huang, G., Guizzardi, D., Koffi, E., Muntean, M., Schieberle, C., Friedrich, R., Janssens-Maenhout, G.: High resolution temporal profiles in the Emissions Database for Global Atmospheric Research, Nature Scientific Data, 2020, in press.

v432: Muntean, M., Guizzardi, D., Schaaf, E., Crippa, M., Solazzo, E., Olivier, J.G.J., Vignati, E. Fossil CO2 emissions of all world countries - 2018 Report, EUR 29433 EN, Publications Office of the European Union, Luxembourg, 2018, ISBN 978-92-79-97240-9, doi:10.2760/30158, JRC113738.

v432_AP:Crippa, M., Guizzardi, D., Muntean, M., Schaaf, E., Dentener, F., van Aardenne, J. A., Monni, S., Doering, U., Olivier, J. G. J., Pagliari, V., and Janssens-Maenhout, G.: Gridded emissions of air pollutants for the period 1970–2012 within EDGAR v4.3.2, Earth Syst. Sci. Data, 10, 1987–2013, https://doi.org/10.5194/essd-10-1987-2018, 2018.

v432_VOC_spec: Huang, G., Brook, R., Crippa, M., Janssens-Maenhout, G., Schieberle, C., Dore, C., Guizzardi, D., Muntean, M., Schaaf, E., and Friedrich, R.: Speciation of anthropogenic emissions of non-methane volatile organic compounds: a global gridded data set for 1970–2012, Atmos. Chem. Phys., 17, 7683-7701, https://doi.org/10.5194/acp-17-7683-2017, 2017.

htap_v2_2: Janssens-Maenhout, G., Crippa, M., Guizzardi, D., Dentener, F., Muntean, M., Pouliot, G., Keating, T., Zhang, Q., Kurokawa, J., Wankmüller, R., Denier van der Gon, H., Kuenen, J. J. P., Klimont, Z., Frost, G., Darras, S., Koffi, B., and Li, M.: HTAP_v2.2: a mosaic of regional and global emission grid maps for 2008 and 2010 to study hemispheric transport of air pollution, Atmos. Chem. Phys., 15, 11411–11432, https://doi.org/10.5194/acp-15-11411-2015, 2015.

MNM is MNN for NOx v432_AP

Examples

if (FALSE) { # \dontrun{
# see all the links:
data(edgar)
head(edgar)
# Download all pollutants for sector
get_edgar(dataset = "v50_AP",
          pol = "CO",
          sector = "TOTALS",
          year = 2014,
          destpath = tempdir(),
          type = "nc",
          ask = FALSE)
} # }