List of WRF emission species

Emission package definitions from WRF 4.0.1, for use in wrf_create function.

data(emis_opt)

Format

A list of emision variables names, same number as emis_opt in namelist.

Source

https://github.com/wrf-model/WRF

Note

look to the number of aerosol of the emis_opt in WRF domumentation / code.

See also

Examples

data(emis_opt)
names(emis_opt)

#>  [1] "no_emission"   "eradm"         "eradmsorg"     "ecbmz_mosaic" 
#>  [5] "ecptec"        "gocart_ecptec" "mozem"         "mozcem"       
#>  [9] "cammam"        "mozmem"        "mozc_t1_em"    "blank1"       
#> [13] "blank2"        "esaprcnov"     "ecb05_opt1"    "ecb05_opt2"   
#> [17] "eco2"          "eghg"          "ecrimech"      "ecrimechtno"  

emis_opt[["eradm"]]

#>  [1] "E_ISO"  "E_SO2"  "E_NO"   "E_CO"   "E_ETH"  "E_HC3"  "E_HC5"  "E_HC8" 
#>  [9] "E_XYL"  "E_OL2"  "E_OLT"  "E_OLI"  "E_TOL"  "E_CSL"  "E_HCHO" "E_ALD" 
#> [17] "E_KET"  "E_ORA2" "E_NH3"