Function to read variables of emission files

Read a variable

wrf_get(
  file = file.choose(),
  name = NA,
  as_raster = FALSE,
  raster_crs = "WRF",
  raster_lev = 1,
  k = NA,
  verbose = FALSE,
  ...
)

Format

array or raster object

Arguments

file

name of file interactively (default) or specified

name

name of the variable (any variable) or time to return a POSIXlt object from model

as_raster

return a raster instead of an array

raster_crs

crs of outputif as_raster is TRUE, see details

raster_lev

level for rasters from a 4D variable

k

multiplier

verbose

display additional information

...

additional parameters passed to wrf_raster

Details

wrf_get can return a raster object with the option as_raster = TRUE, raster_crs can be used to specify the output crs of the raster object, raster_crs = 'latlon' can be especifyed to use latlon option in wrf_raster. If raster_crs is 'WRF' (default), the output projection is equivalent to the WRF grid.

See also

Author

Daniel Schuch

Examples

{

# create the folder and emission file
dir.create(file.path(tempdir(), "EMISS"))
wrf_create(wrfinput_dir = system.file("extdata", package = "eixport"),
         wrfchemi_dir = file.path(tempdir(), "EMISS"))

# get the name of created file
files <- list.files(path = file.path(tempdir(), "EMISS"),
                   pattern = "wrfchemi",
                   full.names = TRUE)

# open, put some numbers and write
CO <- wrf_get(file = files[1],
              name = "E_CO")

CO[] = rnorm(length(CO))

wrf_put(file = files[1],
        name = "E_CO",
        POL = CO)

COr <- wrf_get(file = files[1],
               name = "E_CO",
               as_raster = TRUE)

}
#> Warning: '/tmp/RtmpB6wQOg/EMISS' already exists
#> creating emission for domain 1 ...
#> output file: /tmp/RtmpB6wQOg/EMISS/wrfchemi_d01_2011-08-01_00:00:00