Export emissions to other formats

Export emissions object according to format of file 'Sources.txt' of the model R-LINE

to_rline(
  Emis,
  Z_b,
  Z_e,
  dCL,
  sigmaz0,
  lanes,
  Hw1,
  dw1,
  Hw2,
  dw2,
  Depth,
  Wtop,
  Wbottom,
  experimental = FALSE,
  crs
)

Arguments

Emis
Column with the emissions whose unit must be g/ms.

Z_b
initial meters above sea level (m).

Z_e
final meters above sea level (m).

dCL
offset distance for each source relative to the centerline.

sigmaz0
vertical dispersion (m).

lanes
number of lanes at each street.

Hw1
Height of the barrier 1 (m).

dw1
Distance to barrier 1 (m).

Hw2
height of the barrier 2 (m).

dw2
Distance to barrier 2 (m).

Depth
Depth of the depression. USed for depressed roadway (m).

Wtop
width of the opening at the top of the depression (m).

Wbottom
width of the roadway at the bottom of the depression (m).

experimental
Boolean argument to denote the use of the experimental features (TRUE) or not (FALSE).

crs
Numeric; Coordenade Reference System to project data or not.

Value

Data frame with format for R-LINE model.

Note

Michelle G. Snyder, Akula Venkatram, David K. Heist, Steven G. Perry, William B. Petersen, Vlad Isakov, RLINE: A line source dispersion model for near-surface releases, In Atmospheric Environment, Volume 77, 2013, Pages 748-756, ISSN 1352-2310, https://doi.org/10.1016/j.atmosenv.2013.05.074.

Examples

{
data(emisco)
emisco <- st_explode(emisco)
emisco$V8 <- units::set_units(emisco$V8, "g/ms")
Source <- to_rline(Emis = emisco["V8"],
                   Z_b =0,
                   Z_e =0,
                   dCL = 0,
                   sigmaz0 = 2,
                   lanes = 1)
head(Source)
write.table(x = Source,
            file = paste0(tempdir(), "/Sources.txt"),
            row.names = FALSE,
            sep = " ",
            quote = FALSE)
}

#> Sum: 2090610.32